Activate warning as error on CI

+++ src/widgets/item/freeform.hpp
@@ -52,6 +52,6 @@ private:
     [[nodiscard]] LIBDRAWYWIDGETS_NO_EXPORT QPointF optimizePoint(const QPointF &newPoint);
     std::deque<QPointF> m_currentWindow;
     QPointF m_currentWindowSum{0, 0};
-    int m_bufferSize{7};
+    quint64 m_bufferSize{7};
 };
 LIBDRAWYWIDGETS_EXPORT QDebug operator<<(QDebug d, const FreeformItem &t);
